refactor(avoidance): separate shift line generate process
