+++ planning/behavior_path_planner/include/behavior_path_planner/scene_module/avoidance/avoidance_module.hpp
@@ -19,6 +19,7 @@
 #include "behavior_path_planner/scene_module/scene_module_visitor.hpp"
 #include "behavior_path_planner/utils/avoidance/avoidance_module_data.hpp"
 #include "behavior_path_planner/utils/avoidance/helper.hpp"
+#include "behavior_path_planner/utils/avoidance/shift_line_generator.hpp"
 #include "behavior_path_planner/utils/path_safety_checker/safety_check.hpp"
 
 #include <rclcpp/rclcpp.hpp>
@@ -43,6 +44,8 @@ using motion_utils::calcSignedArcLength;
 using motion_utils::findNearestIndex;
 
 using tier4_planning_msgs::msg::AvoidanceDebugMsg;
+
+using helper::avoidance::AvoidanceHelper;
 
 class AvoidanceModule : public SceneModuleInterface
 {
@@ -221,6 +224,11 @@ private:
    */
   void insertStopPoint(const bool use_constraints_for_decel, ShiftedPath & shifted_path) const;
 
+  /**
+   * @brief insert stop point to return original lane.
+   * @param flag. if it is true, the ego decelerates within accel/jerk constraints.
+   * @param target path.
+   */
   void insertReturnDeadLine(const bool use_constraints_for_decel, ShiftedPath & shifted_path) const;
 
   /**
@@ -289,165 +297,10 @@ private:
   void fillDebugData(const AvoidancePlanningData & data, DebugData & debug) const;
 
   /**
-   * @brief update registered shift lines.
-   * @param current shift lines.
-   */
-  void registerRawShiftLines(const AvoidLineArray & future_registered);
-
-  /**
-   * @brief update path index of the registered objects. remove old objects whose end point is
-   * behind ego pose.
-   */
-  void updateRegisteredRawShiftLines();
-
-  /**
    * @brief check whether the ego can transit yield maneuver.
    * @param avoidance data.
    */
   bool canYieldManeuver(const AvoidancePlanningData & data) const;
-
-  // shift line generation
-
-  /**
-   * @brief Calculate the shift points (start/end point, shift length) from the object lateral
-   * and longitudinal positions in the Frenet coordinate. The jerk limit is also considered here.
-   * @param avoidance data.
-   * @param debug data.
-   * @return processed shift lines.
-   */
-  AvoidOutlines generateAvoidOutline(AvoidancePlanningData & data, DebugData & debug) const;
-
-  /*
-   * @brief merge avoid outlines.
-   * @param original shift lines.
-   * @param debug data.
-   * @return processed shift lines.
-   */
-  AvoidOutlines applyMergeProcess(const AvoidOutlines & outlines, DebugData & debug) const;
-
-  /*
-   * @brief fill gap between two shift lines.
-   * @param original shift lines.
-   * @param debug data.
-   * @return processed shift lines.
-   */
-  AvoidOutlines applyFillGapProcess(const AvoidOutlines & outlines, DebugData & debug) const;
-
-  /*
-   * @brief generate candidate shift lines.
-   * @param one-shot shift lines.
-   * @param debug data.
-   */
-  AvoidLineArray generateCandidateShiftLine(
-    const AvoidLineArray & shift_lines, DebugData & debug) const;
-
-  /**
-   * @brief clean up raw shift lines.
-   * @param target shift lines.
-   * @param debug data.
-   * @return processed shift lines.
-   * @details process flow:
-   * 1. combine raw shirt lines and previous registered shift lines.
-   * 2. add return shift line.
-   * 3. merge raw shirt lines.
-   * 4. trim unnecessary shift lines.
-   */
-  AvoidLineArray applyPreProcess(const AvoidOutlines & outlines, DebugData & debug) const;
-
-  /*
-   * @brief fill gap among shift lines.
-   * @param original shift lines.
-   * @param debug data.
-   * @return processed shift lines.
-   */
-  AvoidLineArray applyFillGapProcess(const AvoidLineArray & shift_lines, DebugData & debug) const;
-
-  /*
-   * @brief merge negative & positive shift lines.
-   * @param original shift lines.
-   * @param debug data.
-   * @return processed shift lines.
-   */
-  AvoidLineArray applyMergeProcess(const AvoidLineArray & shift_lines, DebugData & debug) const;
-
-  /*
-   * @brief add return shift line from ego position.
-   * @param current raw shift line.
-   * @param current registered shift line.
-   * @param debug data.
-   */
-  AvoidLineArray applyCombineProcess(
-    const AvoidLineArray & shift_lines, const AvoidLineArray & registered_lines,
-    [[maybe_unused]] DebugData & debug) const;
-
-  /*
-   * @brief add return shift line from ego position.
-   * @param shift lines which the return shift is added.
-   * Pick up the last shift point, which is the most farthest from ego, from the current candidate
-   * avoidance points and registered points in the shifter. If the last shift length of the point is
-   * non-zero, add a return-shift to center line from the point. If there is no shift point in
-   * candidate avoidance points nor registered points, and base_shift > 0, add a return-shift to
-   * center line from ego.
-   */
-  AvoidLineArray addReturnShiftLine(const AvoidLineArray & shift_lines, DebugData & debug) const;
-
-  /*
-   * @brief extract shift lines from total shift lines based on their gradient.
-   * @param shift length data.
-   * @return extracted shift lines.
-   */
-  AvoidLineArray extractShiftLinesFromLine(ShiftLineData & shift_line_data) const;
-
-  /*
-   * @brief remove unnecessary avoid points
-   * @param original shift lines.
-   * @param debug data.
-   * @return processed shift lines.
-   * @details
-   * - Combine avoid points that have almost same gradient
-   * - Quantize the shift length to reduce the shift point noise
-   * - Change the shift length to the previous one if the deviation is small.
-   * - Remove unnecessary return shift (back to the center line).
-   */
-  AvoidLineArray applyTrimProcess(const AvoidLineArray & shift_lines, DebugData & debug) const;
-
-  /*
-   * @brief extract new shift lines based on current shifted path. the module makes a RTC request
-   * for those new shift lines.
-   * @param candidate shift lines.
-   * @return new shift lines.
-   */
-  AvoidLineArray findNewShiftLine(const AvoidLineArray & shift_lines, DebugData & debug) const;
-
-  /*
-   * @brief generate total shift line. total shift line has shift length and gradient array.
-   * @param raw shift lines.
-   * @param total shift line.
-   */
-  void generateTotalShiftLine(
-    const AvoidLineArray & avoid_points, ShiftLineData & shift_line_data) const;
-
-  /*
-   * @brief quantize shift line length.
-   * @param target shift lines.
-   * @param threshold. shift length is quantized by this value. (if it is 0.3[m], the output shift
-   * length is 0.0, 0.3, 0.6...)
-   */
-  void applyQuantizeProcess(AvoidLineArray & shift_lines, const double threshold) const;
-
-  /*
-   * @brief trim shift line whose relative longitudinal distance is less than threshold.
-   * @param target shift lines.
-   * @param threshold.
-   */
-  void applySmallShiftFilter(AvoidLineArray & shift_lines, const double threshold) const;
-
-  /*
-   * @brief merge multiple shift lines whose relative gradient is less than threshold.
-   * @param target shift lines.
-   * @param threshold.
-   */
-  void applySimilarGradFilter(AvoidLineArray & shift_lines, const double threshold) const;
 
   /**
    * @brief add new shift line to path shifter if the RTC status is activated.
@@ -505,15 +358,6 @@ private:
    */
   bool isSafePath(ShiftedPath & shifted_path, DebugData & debug) const;
 
-  bool isComfortable(const AvoidLineArray & shift_lines) const
-  {
-    return std::all_of(shift_lines.begin(), shift_lines.end(), [&](const auto & line) {
-      return PathShifter::calcJerkFromLatLonDistance(
-               line.getRelativeLength(), line.getRelativeLongitudinal(),
-               helper_.getAvoidanceEgoSpeed()) < helper_.getLateralMaxJerkLimit();
-    });
-  }
-
   // post process
 
   /**
@@ -543,8 +387,7 @@ private:
       removeRTCStatus();
     }
 
-    current_raw_shift_lines_.clear();
-    registered_raw_shift_lines_.clear();
+    generator_.reset();
     path_shifter_.setShiftLines(ShiftLineArray{});
   }
 
@@ -557,15 +400,6 @@ private:
     const size_t idx = planner_data_->findEgoIndex(path_shifter_.getReferencePath().points);
     path_shifter_.removeBehindShiftLineAndSetBaseOffset(idx);
   }
-
-  // misc functions
-
-  double getCurrentBaseShift() const { return path_shifter_.getBaseOffset(); }
-
-  // TODO(Horibe): think later.
-  // for unique ID
-  mutable uint64_t original_unique_id = 0;  // TODO(Horibe) remove mutable
-  uint64_t getOriginalShiftLineUniqueId() const { return original_unique_id++; }
 
   struct RegisteredShiftLine
   {
@@ -582,9 +416,11 @@ private:
 
   bool safe_{true};
 
+  std::shared_ptr<AvoidanceHelper> helper_;
+
   std::shared_ptr<AvoidanceParameters> parameters_;
 
-  helper::avoidance::AvoidanceHelper helper_;
+  utils::avoidance::ShiftLineGenerator generator_;
 
   AvoidancePlanningData avoid_data_;
 
@@ -593,10 +429,6 @@ private:
   RegisteredShiftLineArray left_shift_array_;
 
   RegisteredShiftLineArray right_shift_array_;
-
-  AvoidLineArray registered_raw_shift_lines_;
-
-  AvoidLineArray current_raw_shift_lines_;
 
   UUID candidate_uuid_;
