+++ planning/behavior_path_planner/src/scene_module/avoidance/avoidance_module.cpp
@@ -90,56 +90,6 @@ bool isBestEffort(const std::string & policy)
   return policy == "best_effort";
 }
 
-AvoidLine merge(const AvoidLine & line1, const AvoidLine & line2, const uint64_t id)
-{
-  AvoidLine ret{};
-
-  ret.start_idx = line1.start_idx;
-  ret.start_shift_length = line1.start_shift_length;
-  ret.start_longitudinal = line1.start_longitudinal;
-
-  ret.end_idx = line2.end_idx;
-  ret.end_shift_length = line2.end_shift_length;
-  ret.end_longitudinal = line2.end_longitudinal;
-
-  ret.id = id;
-  ret.object = line1.object;
-
-  return ret;
-}
-
-AvoidLine fill(const AvoidLine & line1, const AvoidLine & line2, const uint64_t id)
-{
-  AvoidLine ret{};
-
-  ret.start_idx = line1.end_idx;
-  ret.start_shift_length = line1.end_shift_length;
-  ret.start_longitudinal = line1.end_longitudinal;
-
-  ret.end_idx = line2.start_idx;
-  ret.end_shift_length = line2.start_shift_length;
-  ret.end_longitudinal = line2.start_longitudinal;
-
-  ret.id = id;
-  ret.object = line1.object;
-
-  return ret;
-}
-
-AvoidLineArray toArray(const AvoidOutlines & outlines)
-{
-  AvoidLineArray ret{};
-  for (const auto & outline : outlines) {
-    ret.push_back(outline.avoid_line);
-    ret.push_back(outline.return_line);
-
-    std::for_each(
-      outline.middle_lines.begin(), outline.middle_lines.end(),
-      [&ret](const auto & line) { ret.push_back(line); });
-  }
-  return ret;
-}
-
 lanelet::BasicLineString3d toLineString3d(const std::vector<Point> & bound)
 {
   lanelet::BasicLineString3d ret{};
@@ -155,8 +105,9 @@ AvoidanceModule::AvoidanceModule(
   std::unordered_map<std::string, std::shared_ptr<ObjectsOfInterestMarkerInterface>> &
     objects_of_interest_marker_interface_ptr_map)
 : SceneModuleInterface{name, node, rtc_interface_ptr_map, objects_of_interest_marker_interface_ptr_map},  // NOLINT
+  helper_{std::make_shared<AvoidanceHelper>(parameters)},
   parameters_{parameters},
-  helper_{parameters}
+  generator_{parameters}
 {
 }
 
@@ -194,12 +145,12 @@ bool AvoidanceModule::canTransitSuccessState()
   const auto & data = avoid_data_;
 
   // Change input lane. -> EXIT.
-  if (!isDrivingSameLane(helper_.getPreviousDrivingLanes(), data.current_lanelets)) {
+  if (!isDrivingSameLane(helper_->getPreviousDrivingLanes(), data.current_lanelets)) {
     RCLCPP_WARN(getLogger(), "Previous module lane is updated. Exit.");
     return true;
   }
 
-  helper_.setPreviousDrivingLanes(data.current_lanelets);
+  helper_->setPreviousDrivingLanes(data.current_lanelets);
 
   // Reach input path end point. -> EXIT.
   {
@@ -248,7 +199,7 @@ bool AvoidanceModule::canTransitSuccessState()
 
   // Be able to canceling avoidance path. -> EXIT.
   if (!has_avoidance_target) {
-    if (!helper_.isShifted() && parameters_->enable_cancel_maneuver) {
+    if (!helper_->isShifted() && parameters_->enable_cancel_maneuver) {
       RCLCPP_INFO(getLogger(), "No objects. Cancel avoidance path. Exit.");
       return true;
     }
@@ -261,7 +212,7 @@ void AvoidanceModule::fillFundamentalData(AvoidancePlanningData & data, DebugDat
 {
   // reference pose
   data.reference_pose =
-    utils::getUnshiftedEgoPose(getEgoPose(), helper_.getPreviousSplineShiftPath());
+    utils::getUnshiftedEgoPose(getEgoPose(), helper_->getPreviousSplineShiftPath());
 
   // lanelet info
   data.current_lanelets = utils::avoidance::getCurrentLanesFromPath(
@@ -288,7 +239,7 @@ void AvoidanceModule::fillFundamentalData(AvoidancePlanningData & data, DebugDat
     parameters_->use_intersection_areas, false));
 
   // reference path
-  if (isDrivingSameLane(helper_.getPreviousDrivingLanes(), data.current_lanelets)) {
+  if (isDrivingSameLane(helper_->getPreviousDrivingLanes(), data.current_lanelets)) {
     data.reference_path_rough = extendBackwardLength(*getPreviousModuleOutput().path);
   } else {
     data.reference_path_rough = *getPreviousModuleOutput().path;
@@ -343,10 +294,10 @@ void AvoidanceModule::fillAvoidanceTargetObjects(
 
   // Separate dynamic objects based on whether they are inside or outside of the expanded lanelets.
   const auto sparse_resample_path = utils::resamplePathWithSpline(
-    helper_.getPreviousSplineShiftPath().path, parameters_->resample_interval_for_output);
+    helper_->getPreviousSplineShiftPath().path, parameters_->resample_interval_for_output);
   const auto [object_within_target_lane, object_outside_target_lane] =
     utils::avoidance::separateObjectsByPath(
-      sparse_resample_path, planner_data_, data, parameters_, helper_.getForwardDetectionRange(),
+      sparse_resample_path, planner_data_, data, parameters_, helper_->getForwardDetectionRange(),
       is_running, debug);
 
   for (const auto & object : object_outside_target_lane.objects) {
@@ -365,7 +316,7 @@ void AvoidanceModule::fillAvoidanceTargetObjects(
   filterTargetObjects(objects, data, debug, planner_data_, parameters_);
 
   // Calculate the distance needed to safely decelerate the ego vehicle to a stop line.
-  const auto feasible_stop_distance = helper_.getFeasibleDecelDistance(0.0, false);
+  const auto feasible_stop_distance = helper_->getFeasibleDecelDistance(0.0, false);
   std::for_each(data.target_objects.begin(), data.target_objects.end(), [&, this](auto & o) {
     o.to_stop_line = calcDistanceToStopLine(o);
     fillObjectStoppableJudge(o, registered_objects_, feasible_stop_distance, parameters_);
@@ -442,7 +393,7 @@ ObjectData AvoidanceModule::createObjectData(
 bool AvoidanceModule::canYieldManeuver(const AvoidancePlanningData & data) const
 {
   // transit yield maneuver only when the avoidance maneuver is not initiated.
-  if (helper_.isShifted()) {
+  if (helper_->isShifted()) {
     RCLCPP_DEBUG(getLogger(), "avoidance maneuver already initiated.");
     return false;
   }
@@ -453,7 +404,7 @@ bool AvoidanceModule::canYieldManeuver(const AvoidancePlanningData & data) const
     const size_t idx = planner_data_->findEgoIndex(path_shifter_.getReferencePath().points);
     const auto to_shift_start_point = calcSignedArcLength(
       path_shifter_.getReferencePath().points, idx, registered_lines.front().start_idx);
-    if (to_shift_start_point < helper_.getMinimumPrepareDistance()) {
+    if (to_shift_start_point < helper_->getMinimumPrepareDistance()) {
       RCLCPP_DEBUG(
         getLogger(),
         "Distance to shift start point is less than minimum prepare distance. The distance is not "
@@ -496,24 +447,13 @@ void AvoidanceModule::fillShiftLine(AvoidancePlanningData & data, DebugData & de
   auto path_shifter = path_shifter_;
 
   /**
-   * STEP1: Generate avoid outlines.
-   * Basically, avoid outlines are generated per target objects.
-   */
-  const auto outlines = generateAvoidOutline(data, debug);
-
-  /**
-   * STEP2: Create rough shift lines.
-   */
-  data.raw_shift_line = applyPreProcess(outlines, debug);
-
-  /**
-   * STEP3: Create candidate shift lines.
+   * STEP1: Create candidate shift lines.
    * Merge rough shift lines, and extract new shift lines.
    */
-  const auto processed_shift_lines = generateCandidateShiftLine(data.raw_shift_line, debug);
+  const auto processed_shift_lines = generator_.generate(data, debug);
 
   /**
-   * Step4: Validate new shift lines.
+   * Step2: Validate new shift lines.
    * Output new shift lines only when the avoidance path which is generated from them doesn't have
    * huge offset from ego.
    */
@@ -524,7 +464,7 @@ void AvoidanceModule::fillShiftLine(AvoidancePlanningData & data, DebugData & de
   data.found_avoidance_path = found_new_sl || registered;
 
   /**
-   * STEP5: Set new shift lines.
+   * STEP3: Set new shift lines.
    * If there are new shift points, these shift points are registered in path_shifter in order to
    * generate candidate avoidance path.
    */
@@ -533,7 +473,7 @@ void AvoidanceModule::fillShiftLine(AvoidancePlanningData & data, DebugData & de
   }
 
   /**
-   * STEP6: Generate avoidance path.
+   * STEP4: Generate avoidance path.
    */
   ShiftedPath spline_shift_path = utils::avoidance::toShiftedPath(data.reference_path);
   const auto success_spline_path_generation =
@@ -543,11 +483,11 @@ void AvoidanceModule::fillShiftLine(AvoidancePlanningData & data, DebugData & de
                           : utils::avoidance::toShiftedPath(data.reference_path);
 
   /**
-   * STEP7: Check avoidance path safety.
+   * STEP5: Check avoidance path safety.
    * For each target objects and the objects in adjacent lanes,
    * check that there is a certain amount of margin in the lateral and longitudinal direction.
    */
-  data.comfortable = isComfortable(data.new_shift_line);
+  data.comfortable = helper_->isComfortable(data.new_shift_line);
   data.safe = isSafePath(data.candidate_path, debug);
 }
 
@@ -590,7 +530,7 @@ void AvoidanceModule::fillEgoStatus(
    */
   if (isOutputPathLocked()) {
     data.safe_shift_line.clear();
-    data.candidate_path = helper_.getPreviousSplineShiftPath();
+    data.candidate_path = helper_->getPreviousSplineShiftPath();
     RCLCPP_DEBUG_THROTTLE(
       getLogger(), *clock_, 500, "this module is locked now. keep current path.");
     return;
@@ -662,7 +602,7 @@ void AvoidanceModule::fillDebugData(
     return;
   }
 
-  if (helper_.isShifted()) {
+  if (helper_->isShifted()) {
     return;
   }
 
@@ -682,9 +622,9 @@ void AvoidanceModule::fillDebugData(
   const auto max_avoid_margin = object_parameter.safety_buffer_lateral * o_front.distance_factor +
                                 object_parameter.avoid_margin_lateral + 0.5 * vehicle_width;
 
-  const auto variable = helper_.getSharpAvoidanceDistance(
-    helper_.getShiftLength(o_front, utils::avoidance::isOnRight(o_front), max_avoid_margin));
-  const auto constant = helper_.getNominalPrepareDistance() +
+  const auto variable = helper_->getSharpAvoidanceDistance(
+    helper_->getShiftLength(o_front, utils::avoidance::isOnRight(o_front), max_avoid_margin));
+  const auto constant = helper_->getNominalPrepareDistance() +
                         object_parameter.safety_buffer_longitudinal +
                         additional_buffer_longitudinal;
   const auto total_avoid_distance = variable + constant;
@@ -754,1203 +694,6 @@ void AvoidanceModule::updateEgoBehavior(const AvoidancePlanningData & data, Shif
   insertReturnDeadLine(isBestEffort(parameters_->policy_deceleration), path);
 
   setStopReason(StopReason::AVOIDANCE, path.path);
-}
-
-void AvoidanceModule::updateRegisteredRawShiftLines()
-{
-  const auto & data = avoid_data_;
-
-  utils::avoidance::fillAdditionalInfoFromPoint(data, registered_raw_shift_lines_);
-
-  AvoidLineArray avoid_lines;
-
-  const auto has_large_offset = [this](const auto & s) {
-    constexpr double THRESHOLD = 0.1;
-    const auto ego_shift_length = helper_.getEgoLinearShift();
-
-    const auto start_to_ego_longitudinal = -1.0 * s.start_longitudinal;
-
-    if (start_to_ego_longitudinal < 0.0) {
-      return false;
-    }
-
-    const auto reg_shift_length =
-      s.getGradient() * start_to_ego_longitudinal + s.start_shift_length;
-
-    return std::abs(ego_shift_length - reg_shift_length) > THRESHOLD;
-  };
-
-  const auto ego_idx = data.ego_closest_path_index;
-
-  for (const auto & s : registered_raw_shift_lines_) {
-    // invalid
-    if (s.end_idx < ego_idx) {
-      continue;
-    }
-
-    // invalid
-    if (has_large_offset(s)) {
-      continue;
-    }
-
-    // valid
-    avoid_lines.push_back(s);
-  }
-
-  DEBUG_PRINT(
-    "ego_closest_path_index = %lu, registered_raw_shift_lines_ size: %lu -> %lu",
-    data.ego_closest_path_index, registered_raw_shift_lines_.size(), avoid_lines.size());
-
-  printShiftLines(registered_raw_shift_lines_, "registered_raw_shift_lines_ (before)");
-  printShiftLines(avoid_lines, "registered_raw_shift_lines_ (after)");
-
-  registered_raw_shift_lines_ = avoid_lines;
-  debug_data_.step1_registered_shift_line = registered_raw_shift_lines_;
-}
-
-AvoidLineArray AvoidanceModule::applyPreProcess(
-  const AvoidOutlines & outlines, DebugData & debug) const
-{
-  AvoidOutlines processed_outlines = outlines;
-
-  /**
-   * Step1: Rough merge process.
-   * Merge multiple avoid outlines. If an avoid outlines' return shift line conflicts other
-   * outline's avoid shift line, those avoid outlines are merged.
-   */
-  processed_outlines = applyMergeProcess(processed_outlines, debug);
-
-  /**
-   * Step2: Fill gap process.
-   * Create and add new shift line to avoid outline in order to fill gaps between avoid shift line
-   * and middle shift lines, return shift line and middle shift lines.
-   */
-  processed_outlines = applyFillGapProcess(processed_outlines, debug);
-
-  /**
-   * Step3: Convert to AvoidLineArray from AvoidOutlines.
-   */
-  AvoidLineArray processed_raw_lines = toArray(processed_outlines);
-
-  /**
-   * Step4: Combine process.
-   * Use all registered points. For the current points, if the similar one of the current
-   * points are already registered, will not use it.
-   */
-  processed_raw_lines =
-    applyCombineProcess(processed_raw_lines, registered_raw_shift_lines_, debug);
-
-  /*
-   * Step5: Add return shift line.
-   * Add return-to-center shift point from the last shift point, if needed.
-   * If there is no shift points, set return-to center shift from ego.
-   */
-  processed_raw_lines = addReturnShiftLine(processed_raw_lines, debug);
-
-  /*
-   * Step6: Fill gap process.
-   * Create and add new shift line to avoid lines.
-   */
-  return applyFillGapProcess(processed_raw_lines, debug);
-}
-
-AvoidLineArray AvoidanceModule::generateCandidateShiftLine(
-  const AvoidLineArray & shift_lines, DebugData & debug) const
-{
-  AvoidLineArray processed_shift_lines = shift_lines;
-
-  /**
-   * Step1: Merge process.
-   * Merge positive shift avoid lines and negative shift avoid lines.
-   */
-  processed_shift_lines = applyMergeProcess(processed_shift_lines, debug);
-
-  /**
-   * Step2: Clean up process.
-   * Remove noisy shift line and concat same gradient shift lines.
-   */
-  processed_shift_lines = applyTrimProcess(processed_shift_lines, debug);
-
-  /**
-   * Step3: Extract new shift lines.
-   * Compare processed shift lines and registered shift lines in order to find new shift lines.
-   */
-  return findNewShiftLine(processed_shift_lines, debug);
-}
-
-void AvoidanceModule::registerRawShiftLines(const AvoidLineArray & future)
-{
-  if (future.empty()) {
-    RCLCPP_ERROR(getLogger(), "future is empty! return.");
-    return;
-  }
-
-  const auto old_size = registered_raw_shift_lines_.size();
-
-  auto future_with_info = future;
-  utils::avoidance::fillAdditionalInfoFromPoint(avoid_data_, future_with_info);
-  printShiftLines(future_with_info, "future_with_info");
-  printShiftLines(registered_raw_shift_lines_, "registered_raw_shift_lines_");
-  printShiftLines(current_raw_shift_lines_, "current_raw_shift_lines_");
-
-  // sort by longitudinal
-  std::sort(future_with_info.begin(), future_with_info.end(), [](auto a, auto b) {
-    return a.end_longitudinal < b.end_longitudinal;
-  });
-
-  // calc relative lateral length
-  future_with_info.front().start_shift_length = getCurrentBaseShift();
-  for (size_t i = 1; i < future_with_info.size(); ++i) {
-    future_with_info.at(i).start_shift_length = future_with_info.at(i - 1).end_shift_length;
-  }
-
-  const auto is_registered = [this](const auto id) {
-    const auto & r = registered_raw_shift_lines_;
-    return std::any_of(r.begin(), r.end(), [id](const auto & s) { return s.id == id; });
-  };
-
-  const auto same_id_shift_line = [this](const auto id) {
-    const auto & r = current_raw_shift_lines_;
-    const auto itr = std::find_if(r.begin(), r.end(), [id](const auto & s) { return s.id == id; });
-    if (itr != r.end()) {
-      return *itr;
-    }
-    throw std::logic_error("not found same id current raw shift line.");
-  };
-
-  for (const auto & s : future_with_info) {
-    if (s.parent_ids.empty()) {
-      RCLCPP_ERROR(getLogger(), "avoid line for path_shifter must have parent_id.");
-    }
-
-    for (const auto id : s.parent_ids) {
-      if (is_registered(id)) {
-        continue;
-      }
-
-      registered_raw_shift_lines_.push_back(same_id_shift_line(id));
-    }
-  }
-
-  DEBUG_PRINT("registered object size: %lu -> %lu", old_size, registered_raw_shift_lines_.size());
-}
-
-AvoidOutlines AvoidanceModule::generateAvoidOutline(
-  AvoidancePlanningData & data, [[maybe_unused]] DebugData & debug) const
-{
-  // To be consistent with changes in the ego position, the current shift length is considered.
-  const auto current_ego_shift = helper_.getEgoShift();
-  const auto & base_link2rear = planner_data_->parameters.base_link2rear;
-
-  // Calculate feasible shift length
-  const auto get_shift_profile =
-    [&](
-      auto & object, const auto & desire_shift_length) -> std::optional<std::pair<double, double>> {
-    // use each object param
-    const auto object_type = utils::getHighestProbLabel(object.object.classification);
-    const auto object_parameter = parameters_->object_parameters.at(object_type);
-    const auto is_object_on_right = utils::avoidance::isOnRight(object);
-
-    // use absolute dist for return-to-center, relative dist from current for avoiding.
-    const auto avoiding_shift = desire_shift_length - current_ego_shift;
-    const auto nominal_avoid_distance = helper_.getMaxAvoidanceDistance(avoiding_shift);
-
-    // calculate remaining distance.
-    const auto prepare_distance = helper_.getNominalPrepareDistance();
-    const auto & additional_buffer_longitudinal =
-      object_parameter.use_conservative_buffer_longitudinal
-        ? planner_data_->parameters.base_link2front
-        : 0.0;
-    const auto constant = object_parameter.safety_buffer_longitudinal +
-                          additional_buffer_longitudinal + prepare_distance;
-    const auto has_enough_distance = object.longitudinal > constant + nominal_avoid_distance;
-    const auto remaining_distance = object.longitudinal - constant;
-    const auto avoidance_distance =
-      has_enough_distance ? nominal_avoid_distance : remaining_distance;
-
-    // nominal case. avoidable.
-    if (has_enough_distance) {
-      return std::make_pair(desire_shift_length, avoidance_distance);
-    }
-
-    if (!isBestEffort(parameters_->policy_lateral_margin)) {
-      return std::make_pair(desire_shift_length, avoidance_distance);
-    }
-
-    // ego already has enough positive shift.
-    const auto has_enough_positive_shift = avoiding_shift < -1e-3 && desire_shift_length > 1e-3;
-    if (is_object_on_right && has_enough_positive_shift) {
-      return std::make_pair(desire_shift_length, avoidance_distance);
-    }
-
-    // ego already has enough negative shift.
-    const auto has_enough_negative_shift = avoiding_shift > 1e-3 && desire_shift_length < -1e-3;
-    if (!is_object_on_right && has_enough_negative_shift) {
-      return std::make_pair(desire_shift_length, avoidance_distance);
-    }
-
-    // don't relax shift length since it can stop in front of the object.
-    if (object.is_stoppable && !parameters_->use_shorten_margin_immediately) {
-      return std::make_pair(desire_shift_length, avoidance_distance);
-    }
-
-    // the avoidance path is already approved
-    const auto & object_pos = object.object.kinematics.initial_pose_with_covariance.pose.position;
-    const auto is_approved = (helper_.getShift(object_pos) > 0.0 && is_object_on_right) ||
-                             (helper_.getShift(object_pos) < 0.0 && !is_object_on_right);
-    if (is_approved) {
-      return std::make_pair(desire_shift_length, avoidance_distance);
-    }
-
-    // prepare distance is not enough. unavoidable.
-    if (remaining_distance < 1e-3) {
-      object.reason = AvoidanceDebugFactor::REMAINING_DISTANCE_LESS_THAN_ZERO;
-      return std::nullopt;
-    }
-
-    // calculate lateral jerk.
-    const auto required_jerk = PathShifter::calcJerkFromLatLonDistance(
-      avoiding_shift, remaining_distance, helper_.getAvoidanceEgoSpeed());
-
-    // relax lateral jerk limit. avoidable.
-    if (required_jerk < helper_.getLateralMaxJerkLimit()) {
-      return std::make_pair(desire_shift_length, avoidance_distance);
-    }
-
-    // avoidance distance is not enough. unavoidable.
-    if (!isBestEffort(parameters_->policy_deceleration)) {
-      object.reason = AvoidanceDebugFactor::TOO_LARGE_JERK;
-      return std::nullopt;
-    }
-
-    // output avoidance path under lateral jerk constraints.
-    const auto feasible_relative_shift_length = PathShifter::calcLateralDistFromJerk(
-      remaining_distance, helper_.getLateralMaxJerkLimit(), helper_.getAvoidanceEgoSpeed());
-
-    if (std::abs(feasible_relative_shift_length) < parameters_->lateral_execution_threshold) {
-      object.reason = "LessThanExecutionThreshold";
-      return std::nullopt;
-    }
-
-    const auto feasible_shift_length =
-      desire_shift_length > 0.0 ? feasible_relative_shift_length + current_ego_shift
-                                : -1.0 * feasible_relative_shift_length + current_ego_shift;
-
-    const auto infeasible =
-      std::abs(feasible_shift_length - object.overhang_dist) <
-      0.5 * planner_data_->parameters.vehicle_width + object_parameter.safety_buffer_lateral;
-    if (infeasible) {
-      RCLCPP_WARN_THROTTLE(
-        getLogger(), *clock_, 1000, "feasible shift length is not enough to avoid. ");
-      object.reason = AvoidanceDebugFactor::TOO_LARGE_JERK;
-      return std::nullopt;
-    }
-
-    {
-      RCLCPP_WARN_THROTTLE(
-        getLogger(), *clock_, 1000, "use feasible shift length. [original: (%.2f) actual: (%.2f)]",
-        std::abs(avoiding_shift), feasible_relative_shift_length);
-    }
-
-    return std::make_pair(feasible_shift_length, avoidance_distance);
-  };
-
-  const auto is_forward_object = [](const auto & object) { return object.longitudinal > 0.0; };
-
-  const auto is_valid_shift_line = [](const auto & s) {
-    return s.start_longitudinal > 0.0 && s.start_longitudinal < s.end_longitudinal;
-  };
-
-  AvoidOutlines outlines;
-  for (auto & o : data.target_objects) {
-    if (!o.avoid_margin.has_value()) {
-      o.reason = AvoidanceDebugFactor::INSUFFICIENT_LATERAL_MARGIN;
-      if (o.avoid_required && is_forward_object(o)) {
-        break;
-      } else {
-        continue;
-      }
-    }
-
-    const auto is_object_on_right = utils::avoidance::isOnRight(o);
-    const auto desire_shift_length =
-      helper_.getShiftLength(o, is_object_on_right, o.avoid_margin.value());
-    if (utils::avoidance::isSameDirectionShift(is_object_on_right, desire_shift_length)) {
-      o.reason = AvoidanceDebugFactor::SAME_DIRECTION_SHIFT;
-      if (o.avoid_required && is_forward_object(o)) {
-        break;
-      } else {
-        continue;
-      }
-    }
-
-    // use each object param
-    const auto object_type = utils::getHighestProbLabel(o.object.classification);
-    const auto object_parameter = parameters_->object_parameters.at(object_type);
-    const auto feasible_shift_profile = get_shift_profile(o, desire_shift_length);
-
-    if (!feasible_shift_profile.has_value()) {
-      if (o.avoid_required && is_forward_object(o)) {
-        break;
-      } else {
-        continue;
-      }
-    }
-
-    // use absolute dist for return-to-center, relative dist from current for avoiding.
-    const auto feasible_return_distance =
-      helper_.getMaxAvoidanceDistance(feasible_shift_profile.value().first);
-
-    AvoidLine al_avoid;
-    {
-      const auto & additional_buffer_longitudinal =
-        object_parameter.use_conservative_buffer_longitudinal
-          ? planner_data_->parameters.base_link2front
-          : 0.0;
-      const auto offset =
-        object_parameter.safety_buffer_longitudinal + additional_buffer_longitudinal;
-      const auto to_shift_end = o.longitudinal - offset;
-      const auto path_front_to_ego =
-        avoid_data_.arclength_from_ego.at(avoid_data_.ego_closest_path_index);
-
-      // start point (use previous linear shift length as start shift length.)
-      al_avoid.start_longitudinal = [&]() {
-        const auto nearest_avoid_distance =
-          std::max(to_shift_end - feasible_shift_profile.value().second, 1e-3);
-
-        if (data.to_start_point > to_shift_end) {
-          return nearest_avoid_distance;
-        }
-
-        const auto minimum_avoid_distance =
-          helper_.getMinAvoidanceDistance(feasible_shift_profile.value().first - current_ego_shift);
-        const auto furthest_avoid_distance = std::max(to_shift_end - minimum_avoid_distance, 1e-3);
-
-        return std::clamp(data.to_start_point, nearest_avoid_distance, furthest_avoid_distance);
-      }();
-
-      al_avoid.start_idx = utils::avoidance::findPathIndexFromArclength(
-        avoid_data_.arclength_from_ego, al_avoid.start_longitudinal + path_front_to_ego);
-      al_avoid.start = avoid_data_.reference_path.points.at(al_avoid.start_idx).point.pose;
-      al_avoid.start_shift_length = helper_.getLinearShift(al_avoid.start.position);
-
-      // end point
-      al_avoid.end_shift_length = feasible_shift_profile.value().first;
-      al_avoid.end_longitudinal = to_shift_end;
-
-      // misc
-      al_avoid.id = getOriginalShiftLineUniqueId();
-      al_avoid.object = o;
-      al_avoid.object_on_right = utils::avoidance::isOnRight(o);
-    }
-
-    AvoidLine al_return;
-    {
-      const auto offset = object_parameter.safety_buffer_longitudinal + base_link2rear + o.length;
-      const auto to_shift_start = o.longitudinal + offset;
-
-      // start point
-      al_return.start_shift_length = feasible_shift_profile.value().first;
-      al_return.start_longitudinal = to_shift_start;
-
-      // end point
-      al_return.end_longitudinal = [&]() {
-        if (data.to_return_point > to_shift_start) {
-          return std::clamp(
-            data.to_return_point, to_shift_start, feasible_return_distance + to_shift_start);
-        }
-
-        return to_shift_start + feasible_return_distance;
-      }();
-      al_return.end_shift_length = 0.0;
-
-      // misc
-      al_return.id = getOriginalShiftLineUniqueId();
-      al_return.object = o;
-      al_return.object_on_right = utils::avoidance::isOnRight(o);
-    }
-
-    if (is_valid_shift_line(al_avoid) && is_valid_shift_line(al_return)) {
-      outlines.emplace_back(al_avoid, al_return);
-    } else {
-      o.reason = "InvalidShiftLine";
-      continue;
-    }
-
-    o.is_avoidable = true;
-  }
-
-  // debug
-  {
-    std::vector<AvoidanceDebugMsg> debug_info_array;
-    const auto append = [&](const auto & o) {
-      AvoidanceDebugMsg debug_info;
-      debug_info.object_id = toHexString(o.object.object_id);
-      debug_info.longitudinal_distance = o.longitudinal;
-      debug_info.lateral_distance_from_centerline = o.lateral;
-      debug_info.allow_avoidance = o.reason == "";
-      debug_info.failed_reason = o.reason;
-      debug_info_array.push_back(debug_info);
-    };
-
-    for (const auto & o : data.target_objects) {
-      append(o);
-    }
-
-    debug_avoidance_initializer_for_shift_line_.clear();
-    debug_avoidance_initializer_for_shift_line_ = std::move(debug_info_array);
-    debug_avoidance_initializer_for_shift_line_time_ = clock_->now();
-  }
-
-  utils::avoidance::fillAdditionalInfoFromLongitudinal(data, outlines);
-
-  debug.step1_current_shift_line = toArray(outlines);
-
-  return outlines;
-}
-
-void AvoidanceModule::generateTotalShiftLine(
-  const AvoidLineArray & avoid_lines, ShiftLineData & shift_line_data) const
-{
-  const auto & path = avoid_data_.reference_path;
-  const auto & arcs = avoid_data_.arclength_from_ego;
-  const auto N = path.points.size();
-
-  auto & sl = shift_line_data;
-
-  sl.shift_line = std::vector<double>(N, 0.0);
-  sl.shift_line_grad = std::vector<double>(N, 0.0);
-
-  sl.pos_shift_line = std::vector<double>(N, 0.0);
-  sl.neg_shift_line = std::vector<double>(N, 0.0);
-
-  sl.pos_shift_line_grad = std::vector<double>(N, 0.0);
-  sl.neg_shift_line_grad = std::vector<double>(N, 0.0);
-
-  // debug
-  sl.shift_line_history = std::vector<std::vector<double>>(avoid_lines.size(), sl.shift_line);
-
-  // take minmax for same directional shift length
-  for (size_t j = 0; j < avoid_lines.size(); ++j) {
-    const auto & al = avoid_lines.at(j);
-    for (size_t i = 0; i < N; ++i) {
-      // calc current interpolated shift
-      const auto i_shift = utils::avoidance::lerpShiftLengthOnArc(arcs.at(i), al);
-
-      // update maximum shift for positive direction
-      if (i_shift > sl.pos_shift_line.at(i)) {
-        sl.pos_shift_line.at(i) = i_shift;
-        sl.pos_shift_line_grad.at(i) = al.getGradient();
-      }
-
-      // update minumum shift for negative direction
-      if (i_shift < sl.neg_shift_line.at(i)) {
-        sl.neg_shift_line.at(i) = i_shift;
-        sl.neg_shift_line_grad.at(i) = al.getGradient();
-      }
-
-      // store for debug print
-      sl.shift_line_history.at(j).at(i) = i_shift;
-    }
-  }
-
-  // Merge shift length of opposite directions.
-  for (size_t i = 0; i < N; ++i) {
-    sl.shift_line.at(i) = sl.pos_shift_line.at(i) + sl.neg_shift_line.at(i);
-    sl.shift_line_grad.at(i) = sl.pos_shift_line_grad.at(i) + sl.neg_shift_line_grad.at(i);
-  }
-
-  // overwrite shift with current_ego_shift until ego pose.
-  const auto current_shift = helper_.getEgoLinearShift();
-  for (size_t i = 0; i < sl.shift_line.size(); ++i) {
-    if (avoid_data_.ego_closest_path_index < i) {
-      break;
-    }
-    sl.shift_line.at(i) = current_shift;
-    sl.shift_line_grad.at(i) = 0.0;
-  }
-
-  // If the shift point does not have an associated object,
-  // use previous value.
-  for (size_t i = 1; i < N; ++i) {
-    bool has_object = false;
-    for (const auto & al : avoid_lines) {
-      if (al.start_idx <= i && i <= al.end_idx) {
-        has_object = true;
-        break;
-      }
-    }
-    if (!has_object) {
-      sl.shift_line.at(i) = sl.shift_line.at(i - 1);
-    }
-  }
-
-  if (avoid_lines.empty()) {
-    sl.shift_line_history.push_back(sl.shift_line);
-    return;
-  }
-
-  const auto grad_first_shift_line = (avoid_lines.front().start_shift_length - current_shift) /
-                                     avoid_lines.front().start_longitudinal;
-
-  for (size_t i = avoid_data_.ego_closest_path_index; i <= avoid_lines.front().start_idx; ++i) {
-    sl.shift_line.at(i) = helper_.getLinearShift(getPoint(path.points.at(i)));
-    sl.shift_line_grad.at(i) = grad_first_shift_line;
-  }
-
-  sl.shift_line_history.push_back(sl.shift_line);
-}
-
-AvoidLineArray AvoidanceModule::extractShiftLinesFromLine(ShiftLineData & shift_line_data) const
-{
-  using utils::avoidance::setEndData;
-  using utils::avoidance::setStartData;
-
-  const auto & path = avoid_data_.reference_path;
-  const auto & arcs = avoid_data_.arclength_from_ego;
-  const auto N = path.points.size();
-
-  auto & sl = shift_line_data;
-
-  const auto backward_grad = [&](const size_t i) {
-    if (i == 0) {
-      return sl.shift_line_grad.at(i);
-    }
-    const double ds = arcs.at(i) - arcs.at(i - 1);
-    if (ds < 1.0e-5) {
-      return sl.shift_line_grad.at(i);
-    }  // use theoretical value when ds is too small.
-    return (sl.shift_line.at(i) - sl.shift_line.at(i - 1)) / ds;
-  };
-
-  const auto forward_grad = [&](const size_t i) {
-    if (i == arcs.size() - 1) {
-      return sl.shift_line_grad.at(i);
-    }
-    const double ds = arcs.at(i + 1) - arcs.at(i);
-    if (ds < 1.0e-5) {
-      return sl.shift_line_grad.at(i);
-    }  // use theoretical value when ds is too small.
-    return (sl.shift_line.at(i + 1) - sl.shift_line.at(i)) / ds;
-  };
-
-  // calculate forward and backward gradient of the shift length.
-  // This will be used for grad-change-point check.
-  sl.forward_grad = std::vector<double>(N, 0.0);
-  sl.backward_grad = std::vector<double>(N, 0.0);
-  for (size_t i = 0; i < N - 1; ++i) {
-    sl.forward_grad.at(i) = forward_grad(i);
-    sl.backward_grad.at(i) = backward_grad(i);
-  }
-
-  AvoidLineArray merged_avoid_lines;
-  AvoidLine al{};
-  bool found_first_start = false;
-  constexpr auto CREATE_SHIFT_GRAD_THR = 0.001;
-  constexpr auto IS_ALREADY_SHIFTING_THR = 0.001;
-  for (size_t i = avoid_data_.ego_closest_path_index; i < N - 1; ++i) {
-    const auto & p = path.points.at(i).point.pose;
-    const auto shift = sl.shift_line.at(i);
-
-    // If the vehicle is already on the avoidance (checked by the first point has shift),
-    // set a start point at the first path point.
-    if (!found_first_start && std::abs(shift) > IS_ALREADY_SHIFTING_THR) {
-      setStartData(al, 0.0, p, i, arcs.at(i));  // start length is overwritten later.
-      found_first_start = true;
-      DEBUG_PRINT("shift (= %f) is not zero at i = %lu. set start shift here.", shift, i);
-    }
-
-    // find the point where the gradient of the shift is changed
-    const bool set_shift_line_flag =
-      std::abs(sl.forward_grad.at(i) - sl.backward_grad.at(i)) > CREATE_SHIFT_GRAD_THR;
-
-    if (!set_shift_line_flag) {
-      continue;
-    }
-
-    if (!found_first_start) {
-      setStartData(al, 0.0, p, i, arcs.at(i));  // start length is overwritten later.
-      found_first_start = true;
-      DEBUG_PRINT("grad change detected. start at i = %lu", i);
-    } else {
-      setEndData(al, shift, p, i, arcs.at(i));
-      al.id = getOriginalShiftLineUniqueId();
-      merged_avoid_lines.push_back(al);
-      setStartData(al, 0.0, p, i, arcs.at(i));  // start length is overwritten later.
-      DEBUG_PRINT("end and start point found at i = %lu", i);
-    }
-  }
-
-  if (std::abs(backward_grad(N - 1)) > CREATE_SHIFT_GRAD_THR) {
-    const auto & p = path.points.at(N - 1).point.pose;
-    const auto shift = sl.shift_line.at(N - 1);
-    setEndData(al, shift, p, N - 1, arcs.at(N - 1));
-    al.id = getOriginalShiftLineUniqueId();
-    merged_avoid_lines.push_back(al);
-  }
-
-  return merged_avoid_lines;
-}
-
-AvoidOutlines AvoidanceModule::applyMergeProcess(
-  const AvoidOutlines & outlines, DebugData & debug) const
-{
-  AvoidOutlines ret{};
-
-  if (outlines.size() < 2) {
-    return outlines;
-  }
-
-  const auto no_conflict = [](const auto & line1, const auto & line2) {
-    return line1.end_idx < line2.start_idx || line2.end_idx < line1.start_idx;
-  };
-
-  const auto same_side_shift = [](const auto & line1, const auto & line2) {
-    return line1.object_on_right == line2.object_on_right;
-  };
-
-  const auto within = [](const auto & line, const size_t idx) {
-    return line.start_idx < idx && idx < line.end_idx;
-  };
-
-  ret.push_back(outlines.front());
-
-  for (size_t i = 1; i < outlines.size(); i++) {
-    auto & last_outline = ret.back();
-    auto & next_outline = outlines.at(i);
-
-    const auto & return_line = last_outline.return_line;
-    const auto & avoid_line = next_outline.avoid_line;
-
-    if (no_conflict(return_line, avoid_line)) {
-      ret.push_back(outlines.at(i));
-      continue;
-    }
-
-    const auto merged_shift_line = merge(return_line, avoid_line, getOriginalShiftLineUniqueId());
-
-    if (!isComfortable(AvoidLineArray{merged_shift_line})) {
-      ret.push_back(outlines.at(i));
-      continue;
-    }
-
-    if (same_side_shift(return_line, avoid_line)) {
-      last_outline.middle_lines.push_back(merged_shift_line);
-      last_outline.return_line = next_outline.return_line;
-      debug.step1_merged_shift_line.push_back(merged_shift_line);
-      continue;
-    }
-
-    if (within(return_line, avoid_line.end_idx) && within(avoid_line, return_line.start_idx)) {
-      last_outline.middle_lines.push_back(merged_shift_line);
-      last_outline.return_line = next_outline.return_line;
-      debug.step1_merged_shift_line.push_back(merged_shift_line);
-      continue;
-    }
-
-    if (within(return_line, avoid_line.start_idx) && within(avoid_line, return_line.end_idx)) {
-      last_outline.middle_lines.push_back(merged_shift_line);
-      last_outline.return_line = next_outline.return_line;
-      debug.step1_merged_shift_line.push_back(merged_shift_line);
-      continue;
-    }
-  }
-
-  utils::avoidance::fillAdditionalInfoFromLongitudinal(avoid_data_, ret);
-  utils::avoidance::fillAdditionalInfoFromLongitudinal(avoid_data_, debug.step1_merged_shift_line);
-
-  return ret;
-}
-
-AvoidOutlines AvoidanceModule::applyFillGapProcess(
-  const AvoidOutlines & outlines, DebugData & debug) const
-{
-  AvoidOutlines ret = outlines;
-
-  for (auto & outline : ret) {
-    if (outline.middle_lines.empty()) {
-      const auto new_line =
-        fill(outline.avoid_line, outline.return_line, getOriginalShiftLineUniqueId());
-      outline.middle_lines.push_back(new_line);
-      debug.step1_filled_shift_line.push_back(new_line);
-    }
-
-    helper_.alignShiftLinesOrder(outline.middle_lines, false);
-
-    if (outline.avoid_line.end_longitudinal < outline.middle_lines.front().start_longitudinal) {
-      const auto new_line =
-        fill(outline.avoid_line, outline.middle_lines.front(), getOriginalShiftLineUniqueId());
-      outline.middle_lines.push_back(new_line);
-      debug.step1_filled_shift_line.push_back(new_line);
-    }
-
-    helper_.alignShiftLinesOrder(outline.middle_lines, false);
-
-    if (outline.middle_lines.back().end_longitudinal < outline.return_line.start_longitudinal) {
-      const auto new_line =
-        fill(outline.middle_lines.back(), outline.return_line, getOriginalShiftLineUniqueId());
-      outline.middle_lines.push_back(new_line);
-      debug.step1_filled_shift_line.push_back(new_line);
-    }
-
-    helper_.alignShiftLinesOrder(outline.middle_lines, false);
-  }
-
-  utils::avoidance::fillAdditionalInfoFromLongitudinal(avoid_data_, ret);
-  utils::avoidance::fillAdditionalInfoFromLongitudinal(avoid_data_, debug.step1_filled_shift_line);
-
-  return ret;
-}
-
-AvoidLineArray AvoidanceModule::applyFillGapProcess(
-  const AvoidLineArray & shift_lines, DebugData & debug) const
-{
-  AvoidLineArray sorted = shift_lines;
-
-  helper_.alignShiftLinesOrder(sorted, false);
-
-  AvoidLineArray ret = sorted;
-
-  if (shift_lines.empty()) {
-    return ret;
-  }
-
-  const auto & data = avoid_data_;
-
-  // fill gap between ego and nearest shift line.
-  if (sorted.front().start_longitudinal > 0.0) {
-    AvoidLine ego_line{};
-    utils::avoidance::setEndData(
-      ego_line, helper_.getEgoLinearShift(), data.reference_pose, data.ego_closest_path_index, 0.0);
-
-    const auto new_line = fill(ego_line, sorted.front(), getOriginalShiftLineUniqueId());
-    ret.push_back(new_line);
-    debug.step1_front_shift_line.push_back(new_line);
-  }
-
-  helper_.alignShiftLinesOrder(sorted, false);
-
-  // fill gap among shift lines.
-  for (size_t i = 0; i < sorted.size() - 1; ++i) {
-    if (sorted.at(i + 1).start_longitudinal < sorted.at(i).end_longitudinal) {
-      continue;
-    }
-
-    const auto new_line = fill(sorted.at(i), sorted.at(i + 1), getOriginalShiftLineUniqueId());
-    ret.push_back(new_line);
-    debug.step1_front_shift_line.push_back(new_line);
-  }
-
-  helper_.alignShiftLinesOrder(ret, false);
-
-  utils::avoidance::fillAdditionalInfoFromLongitudinal(avoid_data_, ret);
-  utils::avoidance::fillAdditionalInfoFromLongitudinal(avoid_data_, debug.step1_front_shift_line);
-
-  return ret;
-}
-
-AvoidLineArray AvoidanceModule::applyCombineProcess(
-  const AvoidLineArray & shift_lines, const AvoidLineArray & registered_lines,
-  DebugData & debug) const
-{
-  debug.step1_registered_shift_line = registered_lines;
-  return utils::avoidance::combineRawShiftLinesWithUniqueCheck(registered_lines, shift_lines);
-}
-
-AvoidLineArray AvoidanceModule::applyMergeProcess(
-  const AvoidLineArray & shift_lines, DebugData & debug) const
-{
-  // Generate shift line by merging shift_lines.
-  ShiftLineData shift_line_data;
-  generateTotalShiftLine(shift_lines, shift_line_data);
-
-  // Re-generate shift points by detecting gradient-change point of the shift line.
-  auto merged_shift_lines = extractShiftLinesFromLine(shift_line_data);
-
-  // set parent id
-  for (auto & al : merged_shift_lines) {
-    al.parent_ids = utils::avoidance::calcParentIds(shift_lines, al);
-  }
-
-  // sort by distance from ego.
-  helper_.alignShiftLinesOrder(merged_shift_lines);
-
-  // debug visualize
-  {
-    debug.pos_shift = shift_line_data.pos_shift_line;
-    debug.neg_shift = shift_line_data.neg_shift_line;
-    debug.total_shift = shift_line_data.shift_line;
-    debug.pos_shift_grad = shift_line_data.pos_shift_line_grad;
-    debug.neg_shift_grad = shift_line_data.neg_shift_line_grad;
-    debug.total_forward_grad = shift_line_data.forward_grad;
-    debug.total_backward_grad = shift_line_data.backward_grad;
-    debug.step2_merged_shift_line = merged_shift_lines;
-  }
-
-  return merged_shift_lines;
-}
-
-AvoidLineArray AvoidanceModule::applyTrimProcess(
-  const AvoidLineArray & shift_lines, DebugData & debug) const
-{
-  if (shift_lines.empty()) {
-    return shift_lines;
-  }
-
-  AvoidLineArray sl_array_trimmed = shift_lines;
-
-  // sort shift points from front to back.
-  helper_.alignShiftLinesOrder(sl_array_trimmed);
-
-  // - Change the shift length to the previous one if the deviation is small.
-  {
-    constexpr double SHIFT_DIFF_THRES = 1.0;
-    applySmallShiftFilter(sl_array_trimmed, SHIFT_DIFF_THRES);
-  }
-
-  // - Combine avoid points that have almost same gradient.
-  // this is to remove the noise.
-  {
-    const auto THRESHOLD = parameters_->same_grad_filter_1_threshold;
-    applySimilarGradFilter(sl_array_trimmed, THRESHOLD);
-    debug.step3_grad_filtered_1st = sl_array_trimmed;
-  }
-
-  // - Quantize the shift length to reduce the shift point noise
-  // This is to remove the noise coming from detection accuracy, interpolation, resampling, etc.
-  {
-    const auto THRESHOLD = parameters_->quantize_filter_threshold;
-    applyQuantizeProcess(sl_array_trimmed, THRESHOLD);
-    debug.step3_quantize_filtered = sl_array_trimmed;
-  }
-
-  // - Change the shift length to the previous one if the deviation is small.
-  {
-    constexpr double SHIFT_DIFF_THRES = 1.0;
-    applySmallShiftFilter(sl_array_trimmed, SHIFT_DIFF_THRES);
-    debug.step3_noise_filtered = sl_array_trimmed;
-  }
-
-  // - Combine avoid points that have almost same gradient (again)
-  {
-    const auto THRESHOLD = parameters_->same_grad_filter_2_threshold;
-    applySimilarGradFilter(sl_array_trimmed, THRESHOLD);
-    debug.step3_grad_filtered_2nd = sl_array_trimmed;
-  }
-
-  // - Combine avoid points that have almost same gradient (again)
-  {
-    const auto THRESHOLD = parameters_->same_grad_filter_3_threshold;
-    applySimilarGradFilter(sl_array_trimmed, THRESHOLD);
-    debug.step3_grad_filtered_3rd = sl_array_trimmed;
-  }
-
-  return sl_array_trimmed;
-}
-
-void AvoidanceModule::applyQuantizeProcess(
-  AvoidLineArray & shift_lines, const double threshold) const
-{
-  if (threshold < 1.0e-5) {
-    return;  // no need to process
-  }
-
-  for (auto & sl : shift_lines) {
-    sl.end_shift_length = std::round(sl.end_shift_length / threshold) * threshold;
-  }
-
-  helper_.alignShiftLinesOrder(shift_lines);
-}
-
-void AvoidanceModule::applySmallShiftFilter(
-  AvoidLineArray & shift_lines, const double threshold) const
-{
-  if (shift_lines.empty()) {
-    return;
-  }
-
-  AvoidLineArray input = shift_lines;
-  shift_lines.clear();
-
-  for (const auto & s : input) {
-    if (s.getRelativeLongitudinal() < threshold) {
-      continue;
-    }
-
-    if (s.start_longitudinal < helper_.getMinimumPrepareDistance()) {
-      continue;
-    }
-
-    shift_lines.push_back(s);
-  }
-}
-
-void AvoidanceModule::applySimilarGradFilter(
-  AvoidLineArray & avoid_lines, const double threshold) const
-{
-  if (avoid_lines.empty()) {
-    return;
-  }
-
-  AvoidLineArray input = avoid_lines;
-  avoid_lines.clear();
-  avoid_lines.push_back(input.front());  // Take the first one anyway (think later)
-
-  AvoidLine base_line = input.front();
-
-  AvoidLineArray combine_buffer;
-  combine_buffer.push_back(input.front());
-
-  constexpr auto SHIFT_THR = 1e-3;
-  const auto is_negative_shift = [&](const auto & s) {
-    return s.getRelativeLength() < -1.0 * SHIFT_THR;
-  };
-
-  const auto is_positive_shift = [&](const auto & s) { return s.getRelativeLength() > SHIFT_THR; };
-
-  for (size_t i = 1; i < input.size(); ++i) {
-    AvoidLine combine{};
-
-    utils::avoidance::setStartData(
-      combine, base_line.start_shift_length, base_line.start, base_line.start_idx,
-      base_line.start_longitudinal);
-    utils::avoidance::setEndData(
-      combine, input.at(i).end_shift_length, input.at(i).end, input.at(i).end_idx,
-      input.at(i).end_longitudinal);
-
-    combine_buffer.push_back(input.at(i));
-
-    const auto violates = [&]() {
-      if (is_negative_shift(input.at(i)) && is_positive_shift(base_line)) {
-        return true;
-      }
-
-      if (is_negative_shift(base_line) && is_positive_shift(input.at(i))) {
-        return true;
-      }
-
-      const auto lon_combine = combine.getRelativeLongitudinal();
-      const auto base_length = base_line.getGradient() * lon_combine;
-      return std::abs(combine.getRelativeLength() - base_length) > threshold;
-    }();
-
-    if (violates) {
-      avoid_lines.push_back(input.at(i));
-      base_line = input.at(i);
-      combine_buffer.clear();
-      combine_buffer.push_back(input.at(i));
-    } else {
-      avoid_lines.back() = combine;
-    }
-  }
-
-  helper_.alignShiftLinesOrder(avoid_lines);
-
-  DEBUG_PRINT("size %lu -> %lu", input.size(), avoid_lines.size());
-}
-
-AvoidLineArray AvoidanceModule::addReturnShiftLine(
-  const AvoidLineArray & shift_lines, DebugData & debug) const
-{
-  AvoidLineArray ret = shift_lines;
-
-  constexpr double ep = 1.0e-3;
-  const auto & data = avoid_data_;
-  const bool has_candidate_point = !ret.empty();
-  const bool has_registered_point = !path_shifter_.getShiftLines().empty();
-
-  const auto exist_unavoidable_object = std::any_of(
-    data.target_objects.begin(), data.target_objects.end(),
-    [](const auto & o) { return !o.is_avoidable && o.longitudinal > 0.0; });
-
-  if (exist_unavoidable_object) {
-    return ret;
-  }
-
-  // If the return-to-center shift points are already registered, do nothing.
-  if (!has_registered_point && std::fabs(getCurrentBaseShift()) < ep) {
-    DEBUG_PRINT("No shift points, not base offset. Do not have to add return-shift.");
-    return ret;
-  }
-
-  constexpr double RETURN_SHIFT_THRESHOLD = 0.1;
-  DEBUG_PRINT("registered last shift = %f", path_shifter_.getLastShiftLength());
-  if (std::abs(path_shifter_.getLastShiftLength()) < RETURN_SHIFT_THRESHOLD) {
-    DEBUG_PRINT("Return shift is already registered. do nothing.");
-    return ret;
-  }
-
-  // From here, the return-to-center is not registered. But perhaps the candidate is
-  // already generated.
-
-  // If it has a shift point, add return shift from the existing last shift point.
-  // If not, add return shift from ego point. (prepare distance is considered for both.)
-  ShiftLine last_sl;  // the return-shift will be generated after the last shift point.
-  {
-    // avoidance points: Yes, shift points: No -> select last avoidance point.
-    if (has_candidate_point && !has_registered_point) {
-      helper_.alignShiftLinesOrder(ret, false);
-      last_sl = ret.back();
-    }
-
-    // avoidance points: No, shift points: Yes -> select last shift point.
-    if (!has_candidate_point && has_registered_point) {
-      last_sl = utils::avoidance::fillAdditionalInfo(
-        data, AvoidLine{path_shifter_.getLastShiftLine().get()});
-    }
-
-    // avoidance points: Yes, shift points: Yes -> select the last one from both.
-    if (has_candidate_point && has_registered_point) {
-      helper_.alignShiftLinesOrder(ret, false);
-      const auto & al = ret.back();
-      const auto & sl = utils::avoidance::fillAdditionalInfo(
-        data, AvoidLine{path_shifter_.getLastShiftLine().get()});
-      last_sl = (sl.end_longitudinal > al.end_longitudinal) ? sl : al;
-    }
-
-    // avoidance points: No, shift points: No -> set the ego position to the last shift point
-    // so that the return-shift will be generated from ego position.
-    if (!has_candidate_point && !has_registered_point) {
-      last_sl.end_idx = avoid_data_.ego_closest_path_index;
-      last_sl.end = avoid_data_.reference_path.points.at(last_sl.end_idx).point.pose;
-      last_sl.end_shift_length = getCurrentBaseShift();
-    }
-  }
-
-  // There already is a shift point candidates to go back to center line, but it could be too sharp
-  // due to detection noise or timing.
-  // Here the return-shift from ego is added for the in case.
-  if (std::fabs(last_sl.end_shift_length) < RETURN_SHIFT_THRESHOLD) {
-    const auto current_base_shift = helper_.getEgoShift();
-    if (std::abs(current_base_shift) < ep) {
-      DEBUG_PRINT("last shift almost is zero, and current base_shift is zero. do nothing.");
-      return ret;
-    }
-
-    // Is there a shift point in the opposite direction of the current_base_shift?
-    //   No  -> we can overwrite the return shift, because the other shift points that decrease
-    //          the shift length are for return-shift.
-    //   Yes -> we can NOT overwrite, because it might be not a return-shift, but a avoiding
-    //          shift to the opposite direction which can not be overwritten by the return-shift.
-    for (const auto & sl : ret) {
-      if (
-        (current_base_shift > 0.0 && sl.end_shift_length < -ep) ||
-        (current_base_shift < 0.0 && sl.end_shift_length > ep)) {
-        DEBUG_PRINT(
-          "try to put overwrite return shift, but there is shift for opposite direction. Skip "
-          "adding return shift.");
-        return ret;
-      }
-    }
-
-    // If return shift already exists in candidate or registered shift lines, skip adding return
-    // shift.
-    if (has_candidate_point || has_registered_point) {
-      return ret;
-    }
-
-    // set the return-shift from ego.
-    DEBUG_PRINT(
-      "return shift already exists, but they are all candidates. Add return shift for overwrite.");
-    last_sl.end_idx = avoid_data_.ego_closest_path_index;
-    last_sl.end = avoid_data_.reference_path.points.at(last_sl.end_idx).point.pose;
-    last_sl.end_shift_length = current_base_shift;
-  }
-
-  const auto & arclength_from_ego = avoid_data_.arclength_from_ego;
-
-  const auto nominal_prepare_distance = helper_.getNominalPrepareDistance();
-  const auto nominal_avoid_distance = helper_.getMaxAvoidanceDistance(last_sl.end_shift_length);
-
-  if (arclength_from_ego.empty()) {
-    return ret;
-  }
-
-  const auto remaining_distance = std::min(
-    arclength_from_ego.back() - parameters_->dead_line_buffer_for_goal,
-    avoid_data_.to_return_point);
-
-  // If the avoidance point has already been set, the return shift must be set after the point.
-  const auto last_sl_distance = avoid_data_.arclength_from_ego.at(last_sl.end_idx);
-
-  // check if there is enough distance for return.
-  if (last_sl_distance > remaining_distance) {  // tmp: add some small number (+1.0)
-    RCLCPP_WARN_THROTTLE(getLogger(), *clock_, 1000, "No enough distance for return.");
-    return ret;
-  }
-
-  // If the remaining distance is not enough, the return shift needs to be shrunk.
-  // (or another option is just to ignore the return-shift.)
-  // But we do not want to change the last shift point, so we will shrink the distance after
-  // the last shift point.
-  //
-  //  The line "===" is fixed, "---" is scaled.
-  //
-  // [Before Scaling]
-  //  ego              last_sl_end             prepare_end            path_end    avoid_end
-  // ==o====================o----------------------o----------------------o------------o
-  //   |            prepare_dist                   |          avoid_dist               |
-  //
-  // [After Scaling]
-  // ==o====================o------------------o--------------------------o
-  //   |        prepare_dist_scaled            |    avoid_dist_scaled     |
-  //
-  const double variable_prepare_distance =
-    std::max(nominal_prepare_distance - last_sl_distance, 0.0);
-
-  double prepare_distance_scaled = std::max(nominal_prepare_distance, last_sl_distance);
-  double avoid_distance_scaled = nominal_avoid_distance;
-  if (remaining_distance < prepare_distance_scaled + avoid_distance_scaled) {
-    const auto scale = (remaining_distance - last_sl_distance) /
-                       std::max(nominal_avoid_distance + variable_prepare_distance, 0.1);
-    prepare_distance_scaled = last_sl_distance + scale * nominal_prepare_distance;
-    avoid_distance_scaled *= scale;
-    DEBUG_PRINT(
-      "last_sl_distance = %f, nominal_prepare_distance = %f, nominal_avoid_distance = %f, "
-      "remaining_distance = %f, variable_prepare_distance = %f, scale = %f, "
-      "prepare_distance_scaled = %f,avoid_distance_scaled = %f",
-      last_sl_distance, nominal_prepare_distance, nominal_avoid_distance, remaining_distance,
-      variable_prepare_distance, scale, prepare_distance_scaled, avoid_distance_scaled);
-  } else {
-    DEBUG_PRINT("there is enough distance. Use nominal for prepare & avoidance.");
-  }
-
-  // shift point for prepare distance: from last shift to return-start point.
-  if (nominal_prepare_distance > last_sl_distance) {
-    AvoidLine al;
-    al.id = getOriginalShiftLineUniqueId();
-    al.start_idx = last_sl.end_idx;
-    al.start = last_sl.end;
-    al.start_longitudinal = arclength_from_ego.at(al.start_idx);
-    al.end_idx =
-      utils::avoidance::findPathIndexFromArclength(arclength_from_ego, prepare_distance_scaled);
-    al.end = avoid_data_.reference_path.points.at(al.end_idx).point.pose;
-    al.end_longitudinal = arclength_from_ego.at(al.end_idx);
-    al.end_shift_length = last_sl.end_shift_length;
-    al.start_shift_length = last_sl.end_shift_length;
-    ret.push_back(al);
-    debug.step1_return_shift_line.push_back(al);
-  }
-
-  // shift point for return to center line
-  {
-    AvoidLine al;
-    al.id = getOriginalShiftLineUniqueId();
-    al.start_idx =
-      utils::avoidance::findPathIndexFromArclength(arclength_from_ego, prepare_distance_scaled);
-    al.start = avoid_data_.reference_path.points.at(al.start_idx).point.pose;
-    al.start_longitudinal = arclength_from_ego.at(al.start_idx);
-    al.end_idx = utils::avoidance::findPathIndexFromArclength(
-      arclength_from_ego, prepare_distance_scaled + avoid_distance_scaled);
-    al.end = avoid_data_.reference_path.points.at(al.end_idx).point.pose;
-    al.end_longitudinal = arclength_from_ego.at(al.end_idx);
-    al.end_shift_length = 0.0;
-    al.start_shift_length = last_sl.end_shift_length;
-    ret.push_back(al);
-    debug.step1_return_shift_line.push_back(al);
-  }
-
-  return ret;
 }
 
 bool AvoidanceModule::isSafePath(
@@ -2048,7 +791,7 @@ bool AvoidanceModule::isSafePath(
 
 PathWithLaneId AvoidanceModule::extendBackwardLength(const PathWithLaneId & original_path) const
 {
-  const auto previous_path = helper_.getPreviousReferencePath();
+  const auto previous_path = helper_->getPreviousReferencePath();
 
   const auto longest_dist_to_shift_point = [&]() {
     double max_dist = 0.0;
@@ -2056,7 +799,7 @@ PathWithLaneId AvoidanceModule::extendBackwardLength(const PathWithLaneId & orig
       max_dist = std::max(
         max_dist, calcSignedArcLength(previous_path.points, pnt.start.position, getEgoPosition()));
     }
-    for (const auto & sp : registered_raw_shift_lines_) {
+    for (const auto & sp : generator_.getRawRegisteredShiftLine()) {
       max_dist = std::max(
         max_dist, calcSignedArcLength(previous_path.points, sp.start.position, getEgoPosition()));
     }
@@ -2128,11 +871,11 @@ BehaviorModuleOutput AvoidanceModule::plan()
 
   // set previous data
   if (success_spline_path_generation && success_linear_path_generation) {
-    helper_.setPreviousLinearShiftPath(linear_shift_path);
-    helper_.setPreviousSplineShiftPath(spline_shift_path);
-    helper_.setPreviousReferencePath(data.reference_path);
+    helper_->setPreviousLinearShiftPath(linear_shift_path);
+    helper_->setPreviousSplineShiftPath(spline_shift_path);
+    helper_->setPreviousReferencePath(data.reference_path);
   } else {
-    spline_shift_path = helper_.getPreviousSplineShiftPath();
+    spline_shift_path = helper_->getPreviousSplineShiftPath();
   }
 
   // post processing
@@ -2168,7 +911,7 @@ BehaviorModuleOutput AvoidanceModule::plan()
     updateDebugMarker(avoid_data_, path_shifter_, debug_data_);
   }
 
-  if (isDrivingSameLane(helper_.getPreviousDrivingLanes(), data.current_lanelets)) {
+  if (isDrivingSameLane(helper_->getPreviousDrivingLanes(), data.current_lanelets)) {
     output.path = std::make_shared<PathWithLaneId>(spline_shift_path.path);
   } else {
     output.path = getPreviousModuleOutput().path;
@@ -2222,14 +965,14 @@ CandidateOutput AvoidanceModule::planCandidate() const
     return output;
   }
 
-  const auto sl = helper_.getMainShiftLine(data.safe_shift_line);
+  const auto sl = helper_->getMainShiftLine(data.safe_shift_line);
   const auto sl_front = data.safe_shift_line.front();
   const auto sl_back = data.safe_shift_line.back();
 
   utils::clipPathLength(
     shifted_path.path, sl_front.start_idx, std::numeric_limits<double>::max(), 0.0);
 
-  output.lateral_shift = helper_.getRelativeShiftToPath(sl);
+  output.lateral_shift = helper_->getRelativeShiftToPath(sl);
   output.start_distance_to_path_change = sl_front.start_longitudinal;
   output.finish_distance_to_path_change = sl_back.end_longitudinal;
 
@@ -2276,17 +1019,15 @@ void AvoidanceModule::updatePathShifter(const AvoidLineArray & shift_lines)
 
   addNewShiftLines(path_shifter_, shift_lines);
 
-  current_raw_shift_lines_ = avoid_data_.raw_shift_line;
+  generator_.setRawRegisteredShiftLine(shift_lines, avoid_data_);
 
-  registerRawShiftLines(shift_lines);
-
-  const auto sl = helper_.getMainShiftLine(shift_lines);
+  const auto sl = helper_->getMainShiftLine(shift_lines);
   const auto sl_front = shift_lines.front();
   const auto sl_back = shift_lines.back();
 
-  if (helper_.getRelativeShiftToPath(sl) > 0.0) {
+  if (helper_->getRelativeShiftToPath(sl) > 0.0) {
     left_shift_array_.push_back({uuid_map_.at("left"), sl_front.start, sl_back.end});
-  } else if (helper_.getRelativeShiftToPath(sl) < 0.0) {
+  } else if (helper_->getRelativeShiftToPath(sl) < 0.0) {
     right_shift_array_.push_back({uuid_map_.at("right"), sl_front.start, sl_back.end});
   }
 
@@ -2354,101 +1095,15 @@ void AvoidanceModule::addNewShiftLines(
   const double road_velocity = avoid_data_.reference_path.points.at(front_new_shift_line.start_idx)
                                  .point.longitudinal_velocity_mps;
   const double shift_time = PathShifter::calcShiftTimeFromJerk(
-    front_new_shift_line.getRelativeLength(), helper_.getLateralMaxJerkLimit(),
-    helper_.getLateralMaxAccelLimit());
+    front_new_shift_line.getRelativeLength(), helper_->getLateralMaxJerkLimit(),
+    helper_->getLateralMaxAccelLimit());
   const double longitudinal_acc =
     std::clamp(road_velocity / shift_time, 0.0, parameters_->max_acceleration);
 
   path_shifter.setShiftLines(future);
   path_shifter.setVelocity(getEgoSpeed());
   path_shifter.setLongitudinalAcceleration(longitudinal_acc);
-  path_shifter.setLateralAccelerationLimit(helper_.getLateralMaxAccelLimit());
-}
-
-AvoidLineArray AvoidanceModule::findNewShiftLine(
-  const AvoidLineArray & shift_lines, DebugData & debug) const
-{
-  if (shift_lines.empty()) {
-    return {};
-  }
-
-  // add small shift lines.
-  const auto add_straight_shift =
-    [&, this](auto & subsequent, bool has_large_shift, const size_t start_idx) {
-      for (size_t i = start_idx; i < shift_lines.size(); ++i) {
-        if (
-          std::abs(shift_lines.at(i).getRelativeLength()) >
-          parameters_->lateral_small_shift_threshold) {
-          if (has_large_shift) {
-            return;
-          }
-
-          has_large_shift = true;
-        }
-
-        if (!isComfortable(AvoidLineArray{shift_lines.at(i)})) {
-          return;
-        }
-
-        subsequent.push_back(shift_lines.at(i));
-      }
-    };
-
-  // get subsequent shift lines.
-  const auto get_subsequent_shift = [&, this](size_t i) {
-    AvoidLineArray subsequent{shift_lines.at(i)};
-
-    if (!isComfortable(subsequent)) {
-      return subsequent;
-    }
-
-    if (shift_lines.size() == i + 1) {
-      return subsequent;
-    }
-
-    if (!isComfortable(AvoidLineArray{shift_lines.at(i + 1)})) {
-      return subsequent;
-    }
-
-    if (
-      std::abs(shift_lines.at(i).getRelativeLength()) <
-      parameters_->lateral_small_shift_threshold) {
-      const auto has_large_shift =
-        shift_lines.at(i + 1).getRelativeLength() > parameters_->lateral_small_shift_threshold;
-
-      // candidate.at(i) is small length shift line. add large length shift line.
-      subsequent.push_back(shift_lines.at(i + 1));
-      add_straight_shift(subsequent, has_large_shift, i + 2);
-    } else {
-      // candidate.at(i) is large length shift line. add small length shift lines.
-      add_straight_shift(subsequent, true, i + 1);
-    }
-
-    return subsequent;
-  };
-
-  // check ignore or not.
-  const auto is_ignore_shift = [this](const auto & s) {
-    return std::abs(helper_.getRelativeShiftToPath(s)) < parameters_->lateral_execution_threshold;
-  };
-
-  for (size_t i = 0; i < shift_lines.size(); ++i) {
-    const auto & candidate = shift_lines.at(i);
-
-    // prevent sudden steering.
-    if (candidate.start_longitudinal < helper_.getMinimumPrepareDistance()) {
-      break;
-    }
-
-    if (!is_ignore_shift(candidate)) {
-      const auto new_shift_lines = get_subsequent_shift(i);
-      debug.step4_new_shift_line = new_shift_lines;
-      return new_shift_lines;
-    }
-  }
-
-  DEBUG_PRINT("No new shift point exists.");
-  return {};
+  path_shifter.setLateralAccelerationLimit(helper_->getLateralMaxAccelLimit());
 }
 
 bool AvoidanceModule::isValidShiftLine(
@@ -2473,7 +1128,7 @@ bool AvoidanceModule::isValidShiftLine(
     const auto new_shift_length = proposed_shift_path.shift_length.at(new_idx);
 
     constexpr double THRESHOLD = 0.1;
-    const auto offset = std::abs(new_shift_length - helper_.getEgoShift());
+    const auto offset = std::abs(new_shift_length - helper_->getEgoShift());
     if (offset > THRESHOLD) {
       RCLCPP_DEBUG_THROTTLE(
         getLogger(), *clock_, 1000, "new shift line is invalid. [HUGE OFFSET (%.2f)]", offset);
@@ -2514,13 +1169,13 @@ void AvoidanceModule::updateData()
 {
   using utils::avoidance::toShiftedPath;
 
-  helper_.setData(planner_data_);
+  helper_->setData(planner_data_);
 
-  if (!helper_.isInitialized()) {
-    helper_.setPreviousSplineShiftPath(toShiftedPath(*getPreviousModuleOutput().path));
-    helper_.setPreviousLinearShiftPath(toShiftedPath(*getPreviousModuleOutput().path));
-    helper_.setPreviousReferencePath(*getPreviousModuleOutput().path);
-    helper_.setPreviousDrivingLanes(utils::avoidance::getCurrentLanesFromPath(
+  if (!helper_->isInitialized()) {
+    helper_->setPreviousSplineShiftPath(toShiftedPath(*getPreviousModuleOutput().path));
+    helper_->setPreviousLinearShiftPath(toShiftedPath(*getPreviousModuleOutput().path));
+    helper_->setPreviousReferencePath(*getPreviousModuleOutput().path);
+    helper_->setPreviousDrivingLanes(utils::avoidance::getCurrentLanesFromPath(
       *getPreviousModuleOutput().reference_path, planner_data_));
   }
 
@@ -2535,8 +1190,11 @@ void AvoidanceModule::updateData()
     return;
   }
 
-  // update registered shift point for new reference path & remove past objects
-  updateRegisteredRawShiftLines();
+  // update shift line generator.
+  generator_.setData(planner_data_);
+  generator_.setPathShifter(path_shifter_);
+  generator_.setHelper(helper_);
+  generator_.update(avoid_data_, debug_data_);
 
   // update shift line and check path safety.
   fillShiftLine(avoid_data_, debug_data_);
@@ -2567,16 +1225,13 @@ void AvoidanceModule::processOnExit()
 
 void AvoidanceModule::initVariables()
 {
-  helper_.reset();
+  helper_->reset();
+  generator_.reset();
   path_shifter_ = PathShifter{};
-
   debug_data_ = DebugData();
   debug_marker_.markers.clear();
   resetPathCandidate();
   resetPathReference();
-  registered_raw_shift_lines_ = {};
-  current_raw_shift_lines_ = {};
-  original_unique_id = 0;
   is_avoidance_maneuver_starts = false;
   arrived_path_end_ = false;
 }
@@ -2594,7 +1249,7 @@ void AvoidanceModule::updateRTCData()
 {
   const auto & data = avoid_data_;
 
-  updateRegisteredRTCStatus(helper_.getPreviousSplineShiftPath().path);
+  updateRegisteredRTCStatus(helper_->getPreviousSplineShiftPath().path);
 
   const auto candidates = data.safe ? data.safe_shift_line : data.new_shift_line;
 
@@ -2603,10 +1258,10 @@ void AvoidanceModule::updateRTCData()
     return;
   }
 
-  const auto shift_line = helper_.getMainShiftLine(candidates);
-  if (helper_.getRelativeShiftToPath(shift_line) > 0.0) {
+  const auto shift_line = helper_->getMainShiftLine(candidates);
+  if (helper_->getRelativeShiftToPath(shift_line) > 0.0) {
     removePreviousRTCStatusRight();
-  } else if (helper_.getRelativeShiftToPath(shift_line) < 0.0) {
+  } else if (helper_->getRelativeShiftToPath(shift_line) < 0.0) {
     removePreviousRTCStatusLeft();
   } else {
     RCLCPP_WARN_STREAM(getLogger(), "Direction is UNKNOWN");
@@ -2618,7 +1273,7 @@ void AvoidanceModule::updateRTCData()
   const auto sl_back = candidates.back();
 
   output.path_candidate = data.candidate_path.path;
-  output.lateral_shift = helper_.getRelativeShiftToPath(shift_line);
+  output.lateral_shift = helper_->getRelativeShiftToPath(shift_line);
   output.start_distance_to_path_change = sl_front.start_longitudinal;
   output.finish_distance_to_path_change = sl_back.end_longitudinal;
 
@@ -2636,7 +1291,7 @@ TurnSignalInfo AvoidanceModule::calcTurnSignalInfo(const ShiftedPath & path) con
   const size_t start_idx = front_shift_line.start_idx;
   const size_t end_idx = front_shift_line.end_idx;
 
-  const auto current_shift_length = helper_.getEgoShift();
+  const auto current_shift_length = helper_->getEgoShift();
   const double start_shift_length = path.shift_length.at(start_idx);
   const double end_shift_length = path.shift_length.at(end_idx);
   const double segment_shift_length = end_shift_length - start_shift_length;
@@ -2846,7 +1501,8 @@ void AvoidanceModule::updateDebugMarker(
   // registering process
   {
     addShiftLine(shifter.getShiftLines(), "step4_old_shift_line", 1.0, 1.0, 0.0, 0.3);
-    addAvoidLine(data.raw_shift_line, "step4_raw_shift_line", 1.0, 0.0, 0.0, 0.3);
+    addAvoidLine(generator_.getRawShiftLine(), "step4_raw_shift_line", 1.0, 0.0, 0.0, 0.3);
+    // addAvoidLine(data.raw_shift_line, "step4_raw_shift_line", 1.0, 0.0, 0.0, 0.3);
     addAvoidLine(data.new_shift_line, "step4_new_shift_line", 1.0, 0.0, 0.0, 0.3);
   }
 
@@ -2926,8 +1582,8 @@ double AvoidanceModule::calcDistanceToStopLine(const ObjectData & object) const
 
   const auto avoid_margin = object_parameter.safety_buffer_lateral * object.distance_factor +
                             object_parameter.avoid_margin_lateral + 0.5 * vehicle_width;
-  const auto variable = helper_.getMinAvoidanceDistance(
-    helper_.getShiftLength(object, utils::avoidance::isOnRight(object), avoid_margin));
+  const auto variable = helper_->getMinAvoidanceDistance(
+    helper_->getShiftLength(object, utils::avoidance::isOnRight(object), avoid_margin));
   const auto & additional_buffer_longitudinal =
     object_parameter.use_conservative_buffer_longitudinal
       ? planner_data_->parameters.base_link2front
@@ -2962,7 +1618,7 @@ void AvoidanceModule::insertReturnDeadLine(
     shifted_path.path.points, getEgoPosition(), shifted_path.path.points.size() - 1);
   const auto buffer = std::max(0.0, to_shifted_path_end - to_reference_path_end);
 
-  const auto min_return_distance = helper_.getMinAvoidanceDistance(shift_length);
+  const auto min_return_distance = helper_->getMinAvoidanceDistance(shift_length);
   const auto to_stop_line = data.to_return_point - min_return_distance - buffer;
 
   // If we don't need to consider deceleration constraints, insert a deceleration point
@@ -2975,7 +1631,7 @@ void AvoidanceModule::insertReturnDeadLine(
   }
 
   // If the stop distance is not enough for comfortable stop, don't insert wait point.
-  const auto is_comfortable_stop = helper_.getFeasibleDecelDistance(0.0) < to_stop_line;
+  const auto is_comfortable_stop = helper_->getFeasibleDecelDistance(0.0) < to_stop_line;
   if (!is_comfortable_stop) {
     RCLCPP_DEBUG(getLogger(), "stop distance is not enough.");
     return;
@@ -2986,7 +1642,7 @@ void AvoidanceModule::insertReturnDeadLine(
 
   // insert slow down speed.
   const double current_target_velocity = PathShifter::calcFeasibleVelocityFromJerk(
-    shift_length, helper_.getLateralMinJerkLimit(), to_stop_line);
+    shift_length, helper_->getLateralMinJerkLimit(), to_stop_line);
   if (current_target_velocity < getEgoSpeed()) {
     RCLCPP_DEBUG(getLogger(), "current velocity exceeds target slow down speed.");
     return;
@@ -3004,7 +1660,7 @@ void AvoidanceModule::insertReturnDeadLine(
 
     // target speed with nominal jerk limits.
     const double v_target = PathShifter::calcFeasibleVelocityFromJerk(
-      shift_length, helper_.getLateralMinJerkLimit(), shift_longitudinal_distance);
+      shift_length, helper_->getLateralMinJerkLimit(), shift_longitudinal_distance);
     const double v_original = shifted_path.path.points.at(i).point.longitudinal_velocity_mps;
     const double v_insert =
       std::max(v_target - parameters_->buf_slow_down_speed, parameters_->min_slow_down_speed);
@@ -3022,7 +1678,7 @@ void AvoidanceModule::insertWaitPoint(
     return;
   }
 
-  if (helper_.isShifted()) {
+  if (helper_->isShifted()) {
     return;
   }
 
@@ -3035,7 +1691,7 @@ void AvoidanceModule::insertWaitPoint(
   }
 
   // If the stop distance is not enough for comfortable stop, don't insert wait point.
-  const auto is_comfortable_stop = helper_.getFeasibleDecelDistance(0.0) < data.to_stop_line;
+  const auto is_comfortable_stop = helper_->getFeasibleDecelDistance(0.0) < data.to_stop_line;
   const auto is_slow_speed = getEgoSpeed() < parameters_->min_slow_down_speed;
   if (!is_comfortable_stop && !is_slow_speed) {
     RCLCPP_WARN_THROTTLE(getLogger(), *clock_, 500, "not execute uncomfortable deceleration.");
@@ -3051,7 +1707,7 @@ void AvoidanceModule::insertWaitPoint(
 
   // If the object cannot be stopped for, calculate a "mild" deceleration distance
   // and insert a deceleration point at that distance
-  const auto stop_distance = helper_.getFeasibleDecelDistance(0.0, false);
+  const auto stop_distance = helper_->getFeasibleDecelDistance(0.0, false);
   utils::avoidance::insertDecelPoint(
     getEgoPosition(), stop_distance, 0.0, shifted_path.path, stop_pose_);
 }
@@ -3094,7 +1750,7 @@ void AvoidanceModule::insertStopPoint(
   }
 
   // Otherwise, consider deceleration constraints before inserting deceleration point
-  const auto decel_distance = helper_.getFeasibleDecelDistance(0.0, false);
+  const auto decel_distance = helper_->getFeasibleDecelDistance(0.0, false);
   if (stop_distance < decel_distance) {
     return;
   }
@@ -3113,11 +1769,11 @@ void AvoidanceModule::insertYieldVelocity(ShiftedPath & shifted_path) const
     return;
   }
 
-  if (helper_.isShifted()) {
+  if (helper_->isShifted()) {
     return;
   }
 
-  const auto decel_distance = helper_.getFeasibleDecelDistance(p->yield_velocity, false);
+  const auto decel_distance = helper_->getFeasibleDecelDistance(p->yield_velocity, false);
   utils::avoidance::insertDecelPoint(
     getEgoPosition(), decel_distance, p->yield_velocity, shifted_path.path, slow_pose_);
 }
@@ -3132,7 +1788,7 @@ void AvoidanceModule::insertPrepareVelocity(ShiftedPath & shifted_path) const
   }
 
   // insert slow down speed only when the avoidance maneuver is not initiated.
-  if (helper_.isShifted()) {
+  if (helper_->isShifted()) {
     return;
   }
 
@@ -3156,13 +1812,13 @@ void AvoidanceModule::insertPrepareVelocity(ShiftedPath & shifted_path) const
   const auto avoid_margin = object_parameter.safety_buffer_lateral * object.distance_factor +
                             object_parameter.avoid_margin_lateral + 0.5 * vehicle_width;
   const auto shift_length =
-    helper_.getShiftLength(object, utils::avoidance::isOnRight(object), avoid_margin);
+    helper_->getShiftLength(object, utils::avoidance::isOnRight(object), avoid_margin);
 
   // check slow down feasibility
-  const auto min_avoid_distance = helper_.getMinAvoidanceDistance(shift_length);
+  const auto min_avoid_distance = helper_->getMinAvoidanceDistance(shift_length);
   const auto distance_to_object = object.longitudinal;
   const auto remaining_distance = distance_to_object - min_avoid_distance;
-  const auto decel_distance = helper_.getFeasibleDecelDistance(parameters_->velocity_map.front());
+  const auto decel_distance = helper_->getFeasibleDecelDistance(parameters_->velocity_map.front());
   if (remaining_distance < decel_distance) {
     return;
   }
@@ -3172,7 +1828,7 @@ void AvoidanceModule::insertPrepareVelocity(ShiftedPath & shifted_path) const
 
   // insert slow down speed.
   const double current_target_velocity = PathShifter::calcFeasibleVelocityFromJerk(
-    shift_length, helper_.getLateralMinJerkLimit(), distance_to_object);
+    shift_length, helper_->getLateralMinJerkLimit(), distance_to_object);
   if (current_target_velocity < getEgoSpeed() && decel_distance < remaining_distance) {
     utils::avoidance::insertDecelPoint(
       getEgoPosition(), decel_distance, parameters_->velocity_map.front(), shifted_path.path,
@@ -3192,7 +1848,7 @@ void AvoidanceModule::insertPrepareVelocity(ShiftedPath & shifted_path) const
 
     // target speed with nominal jerk limits.
     const double v_target = PathShifter::calcFeasibleVelocityFromJerk(
-      shift_length, helper_.getLateralMinJerkLimit(), shift_longitudinal_distance);
+      shift_length, helper_->getLateralMinJerkLimit(), shift_longitudinal_distance);
     const double v_original = shifted_path.path.points.at(i).point.longitudinal_velocity_mps;
     const double v_insert = std::max(v_target - parameters_->buf_slow_down_speed, lower_speed);
