+++ planning/behavior_path_planner/src/utils/avoidance/utils.cpp
@@ -741,22 +741,24 @@ size_t findPathIndexFromArclength(
   return path_arclength_arr.size() - 1;
 }
 
-std::vector<size_t> concatParentIds(
-  const std::vector<size_t> & ids1, const std::vector<size_t> & ids2)
+std::vector<UUID> concatParentIds(const std::vector<UUID> & ids1, const std::vector<UUID> & ids2)
 {
-  std::set<size_t> id_set{ids1.begin(), ids1.end()};
-  for (const auto id : ids2) {
-    id_set.insert(id);
+  std::vector<UUID> ret;
+  for (const auto id2 : ids2) {
+    if (std::any_of(ids1.begin(), ids1.end(), [&id2](const auto & id1) { return id1 == id2; })) {
+      continue;
+    }
+    ret.push_back(id2);
   }
-  const auto v = std::vector<size_t>{id_set.begin(), id_set.end()};
-  return v;
+
+  return ret;
 }
 
-std::vector<size_t> calcParentIds(const AvoidLineArray & lines1, const AvoidLine & lines2)
+std::vector<UUID> calcParentIds(const AvoidLineArray & lines1, const AvoidLine & lines2)
 {
   // Get the ID of the original AP whose transition area overlaps with the given AP,
   // and set it to the parent id.
-  std::set<uint64_t> ids;
+  std::vector<UUID> ret;
   for (const auto & al : lines1) {
     const auto p_s = al.start_longitudinal;
     const auto p_e = al.end_longitudinal;
@@ -766,9 +768,9 @@ std::vector<size_t> calcParentIds(const AvoidLineArray & lines1, const AvoidLine
       continue;
     }
 
-    ids.insert(al.id);
+    ret.push_back(al.id);
   }
-  return std::vector<size_t>(ids.begin(), ids.end());
+  return ret;
 }
 
 double lerpShiftLengthOnArc(double arc, const AvoidLine & ap)
